fix: sign extension
Unsigned values promoted to signed types, implicitly.
It would be possible to get the values sign extended,
which could lead to wrong results.

Explictly casting the values to wider unsigned types
to avoid issues.

+++ src/api.c
@@ -845,7 +845,7 @@ struct ffa_value api_ffa_partition_info_get_regs(struct vcpu *current,
 
 	ret.func = FFA_SUCCESS_64;
 	ret.arg2 = (sizeof(struct ffa_partition_info) & 0xFFFF) << 48;
-	ret.arg2 |= curr_idx << 16;
+	ret.arg2 |= ((uint64_t)curr_idx) << 16;
 	ret.arg2 |= max_idx;
 
 	if (num_entries_to_ret > 1) {
@@ -981,7 +981,8 @@ struct ffa_value ffa_msg_recv_return(const struct vm *receiver)
 	case FFA_MSG_SEND_32:
 		return (struct ffa_value){
 			.func = FFA_MSG_SEND_32,
-			.arg1 = (receiver->mailbox.recv_sender << 16) |
+			.arg1 = ((uint64_t)(receiver->mailbox.recv_sender)
+				 << 16) |
 				receiver->id,
 			.arg3 = receiver->mailbox.recv_size};
 	default:
